asdasd

+++ Angular/proyectoFinal/src/app/paginas/ranking-page/ranking-page.component.ts
@@ -18,10 +18,12 @@ export class RankingPageComponent implements OnInit {
   teacher?:TeachersData;
   return:any;
   new_points:number;
+  name_practica:string;
   constructor(private route: ActivatedRoute,private rankingService: RankingService,private teacherService: TeacherService) {
     this.rankingId=0;
     this.teacher=this.teacherService.teacher;
     this.new_points=0;
+    this.name_practica="";
   }
   
   ngOnInit() {
@@ -38,25 +40,28 @@ export class RankingPageComponent implements OnInit {
   }
 
   eliminarRegistro(id_rank:number,id_student:number){
-    console.log(id_rank,id_student)
-    this.rankingService.deleteStudenRanking(id_rank,id_student).subscribe({
-      
-      
-      next: (value:any)=>{
-        this.return=value
-      }
+    if (confirm('¿Estás seguro de que quieres eliminar este registro?')) {
+      console.log(id_rank,id_student)
+      this.rankingService.deleteStudenRanking(id_rank,id_student).subscribe({
+        next: (value:any)=>{
+          this.return=value
+        }
+      });
+      this.rankingService.getRankingAnalysis(id_rank).subscribe(data => {
+        this.rankingAnalises=data;
+      });
     }
-      );
-     
-    
-    this.rankingService.getRankingAnalysis(id_rank).subscribe(data => {
-      this.rankingAnalises=data;});
-      alert()
+  }
+
+  nuevaPractica(){
+
   }
 
   editarPuntos(id_rank:number,id_student:number){
     console.log(this.new_points)
-   // this.rankingService.editPointStuden(id_student,id_rank,this.new_points).subscribe();
+   this.rankingService.editPointStuden(id_student,id_rank,this.new_points).subscribe(
+   
+   );
 
   }
 }
